# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Сервис Origummy",
        "description": "Сервис для магазина Origummy",
        "contact": {
            "email": "asekenskigo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/origummy"
        }
    ],
    "paths": {
...
        "/v1/promotions/all": {
            "get": {
                "tags": [
                    "v1",
                    "admin",
                    "promotion"
                ],
                "summary": "Список всех акций",
                "description": "Список всех акций",
                "operationId": "getAllPromotions",
                "responses": {
                    "200": {
                        "description": "Успешно получены"
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            }
        },
        "/v1/promotions": {
            "get": {
                "tags": [
                    "v1",
                    "admin",
                    "promotion"
                ],
                "summary": "Список акций",
                "description": "Список акций",
                "operationId": "getPromotions",
                "responses": {
                    "200": {
                        "description": "Успешно получены"
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            },
            "post": {
                "tags": [
                    "v1",
                    "admin",
                    "promotion"
                ],
                "summary": "Создать акцию",
                "description": "Создать акцию",
                "operationId": "createPromotion",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PromotionCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Акция успешно создана"
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            }
        },
        "/v1/promotions/search": {
            "get": {
                "tags": [
                    "v1",
                    "admin",
                    "promotion"
                ],
                "summary": "Поиск по названию акций",
                "description": "Поиск по названию акций",
                "operationId": "promotionSearch",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Value"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно получены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromotionResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            }
        },
        "/v1/promotions/{promotion}": {
            "get": {
                "tags": [
                    "v1",
                    "admin",
                    "promotion"
                ],
                "summary": "Получить акцию по ID",
                "description": "Получить акцию по ID",
                "operationId": "getPromotionById",
                "responses": {
                    "200": {
                        "description": "Данные по акций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromotionResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            },
            "put": {
                "tags": [
                    "v1",
                    "admin",
                    "promotion"
                ],
                "summary": "Изменить акцию",
                "description": "Изменить акцию",
                "operationId": "updatePromotion",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PromotionUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Акция успешно изменена"
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            },
            "delete": {
                "tags": [
                    "v1",
                    "admin",
                    "promotion"
                ],
                "summary": "Удалить акцию",
                "description": "Удалить акцию",
                "operationId": "deletePromotion",
                "responses": {
                    "200": {
                        "description": "Акция успешно удалена"
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            }
        },
        "/v1/promotions/types": {
            "get": {
                "tags": [
                    "v1",
                    "admin",
                    "promotion"
                ],
                "summary": "Получить типы акций",
                "description": "Получить типы акций",
                "operationId": "getPromotionTypes",
                "responses": {
                    "200": {
                        "description": "Успешно получены"
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PromotionCreateRequest": {
                "required": [
                    "title",
                    "type_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "type_id": {
                        "type": "integer"
                    },
                    "related_products": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
            "PromotionUpdateRequest": {
                "required": [
                    "id",
                    "type_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "type_id": {
                        "type": "integer"
                    },
                    "related_products": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "delete_related_products": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
...
            "PromotionResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "type_id": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "object"
                    },
                    "relatedProducts": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}